Extract most code from `define_feedable!`
